added new handler.go file
